refactor(node): improve cpu and ram calculation accuracy
- Change NodeUsage fields from float/long to double/long for better precision
- Rewrite cpu calculation logic to handle edge cases and improve accuracy
- Simplify null checks using Optional
- Clean up stream completion and error handling

+++ server/src/main/java/server/types/data/NodeUsage.java
@@ -2,5 +2,5 @@ package server.types.data;
 
 import java.time.Instant;
 
-public record NodeUsage(float cpu, float ram, Instant createdAt) {
+public record NodeUsage(double cpu, long ram, Instant createdAt) {
 }
